separate the menu into its own component
