+++ src/components/Menu.js
@@ -1,42 +1,38 @@
-import React from 'react';
-import Styled from 'styled-components';
+import React, { useState } from 'react';
 import { NavLink } from 'react-router-dom';
 
-const Menu = ({ auth }) => {
+// STYLES
+import {
+  MenuToggle,
+  SideMenu
+} from '../styles/menu.css';
+
+const Menu = () => {
+  const [menu, setMenu] = useState(false);
+
+  const toggleMenu = () => setMenu(!menu);
+
   return (
-    <MenuContainer>
-      <NavItem>Menu</NavItem>
-      <MenuNavLink to="/games">Games</MenuNavLink>
-      <MenuNavLink to="/billing">Billing</MenuNavLink>
-      <MenuNavLink to="/profile">Settings</MenuNavLink>
-    </MenuContainer>
+    <nav>
+      <MenuToggle onClick={toggleMenu} menu={menu}>
+        <div className="bar" />
+        <div className="bar" />
+        <div className="bar" />
+      </MenuToggle>
+      <SideMenu menu={menu}>
+        <h1>TVZ</h1>
+        <div>
+          <NavLink to="/games">Games</NavLink>
+        </div>
+        <div>
+          <NavLink to="/billing">Billing</NavLink>
+        </div>
+        <div>
+          <NavLink to="/settings">Settings</NavLink>
+        </div>
+      </SideMenu>
+    </nav>
   );
 };
-
-const MenuContainer = Styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: center;
-  background-color: rgba(31, 71, 115, 1);
-  padding: 3vh 5vw;
-  margin: 5vh 0;
-  width: 40vw;
-`;
-
-const NavItem = Styled.div`
-  display: flex;
-  flex-direction: row;
-  color: white;
-  text-decoration: none;
-`;
-
-const MenuNavLink = Styled(NavLink)`
-  display: flex;
-  flex-direction: row;
-  cursor: pointer;
-  color: white;
-  text-decoration: none;
-`;
 
 export default Menu;
